# developer-docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Persona.ai Developer Platform API",
    "version": "1.0.0",
    "description": "REST + AG-UI API for integrating your product with Persona's agent infrastructure.\n\nEvery endpoint here is a server-to-server call — see the Integration Guide for where in your own stack to call each one from, and how the same endpoint serves both Project-level and per-end-user requests via the `x-persona-external-user-id` header."
  },
  "servers": [
    {
      "url": "https://api.persona.hasanraiyan.me",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "projectCredential": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "<keyId>.<secret>",
        "description": "Developer Platform Project credential, minted via POST /api/v1/projects/{projectId}/credentials. Send as `Authorization: Bearer <keyId>.<secret>`. Optionally pair with the `x-persona-external-user-id` header to act on behalf of one of the Project's own external users. This credential is a server-side secret — see the Integration Guide."
      }
    },
...
  },
  "paths": {
    "/api/v1/developer/threads": {
      "post": {
        "tags": [
          "Developer"
        ],
        "summary": "Create a Thread for the asserted external user",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "x-persona-external-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "agentId"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Thread created"
          },
          "400": {
            "description": "Validation error, or no asserted external user"
          },
          "404": {
            "description": "Agent not found (or not executable by this Domain/Subject)"
          }
        }
      },
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "List the asserted external user's own Threads",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "x-persona-external-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of Threads"
          }
        }
      }
    },
    "/api/v1/developer/threads/{threadId}": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Get a Thread",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-persona-external-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thread"
          },
          "404": {
            "description": "Thread not found (or not visible to this Domain/Subject)"
          }
        }
      },
      "patch": {
        "tags": [
          "Developer"
        ],
        "summary": "Update a Thread's title and/or archived status (Subject only)",
        "description": "Both fields are optional — send just `{ isArchived: true }` to archive a Thread without touching its title, or just `{ title }` to rename it, or both together.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "isArchived": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Thread"
          },
          "404": {
            "description": "Thread not found or unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Developer"
        ],
        "summary": "Delete a Thread (Subject only)",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Thread deleted"
          },
          "404": {
            "description": "Thread not found or unauthorized"
          }
        }
      }
    },
    "/api/v1/developer/threads/bulk-delete": {
      "post": {
        "tags": [
          "Developer"
        ],
        "summary": "Delete multiple of the asserted external user's own Threads in one call",
        "description": "Best-effort batch delete — one not-found id doesn't abort the rest. Up to 100 ids per request.\n",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "x-persona-external-user-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ids"
                ],
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ deleted: string[], failed: [{ id, reason }] }"
          },
          "400": {
            "description": "ids missing, empty, over 100 entries, or no asserted external user"
          }
        }
      }
    },
    "/api/v1/developer/threads/{threadId}/messages": {
      "get": {
        "tags": [
          "Developer"
        ],
        "summary": "Get a Thread's message history",
        "security": [
          {
            "projectCredential": []
          }
        ],
        "parameters": [
          {
            "name": "threadId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message list with subagent traces"
          },
          "404": {
            "description": "Thread not found or unauthorized"
          }
        }
      }
    },
...
  },
...
}